refactor tibber api
